requirements.txt
Pillow>=10.0.0
python-dotenv>=1.0.1
elevenlabs>=1.2.0
Flask>=3.0.3
langchain>=0.2.5
langchain-core>=0.2.5
langchain-openai>=0.1.7
mutagen>=1.47.0
openai>=1.40.0
requests>=2.31.0